Rename normalize/normalize.rs to normalizer.rs
Rename normalize/normalize.rs to normalize/normalizer.rs to avoid
"module inception" warnings and for the sake of consistency with
inspection stuff.
